+++ .gitignore
@@ -9,12 +9,12 @@ __pycache__/
 build/
 dist/
 
-# Aufgabeneditor
+# Task Editor
 # --------------------
 task_editor/credentials.json
 task_editor/temp_repo_*/
 
-# Programmlogik
+# Program Logic
 # --------------------
 program_logic/json_cache.json
 # local cache, not intented to be shared, as it is generated when initializing the program and can be easily recreated
