--- Dtos/__init__.py
@@ -1,15 +1,0 @@
-from Dtos.catalogue_dto import CatalogueDto
-from Dtos.found_task_dto import FoundTaskDto
-from Dtos.subfield_dto import SubfieldDto
-from Dtos.task_dto import TaskDto
-from Dtos.field_dto import FieldDto
-from Dtos.credentials_dto import ConfigDto
-
-__all__ = [
-	"CatalogueDto",
-	"FoundTaskDto",
-	"SubfieldDto",
-	"TaskDto",
-	"FieldDto",
-    "ConfigDto",
-]
