+++ documentation/Konzeption_Editor_GUI.png
