+++ documentation/Konzeption_Farbwahl_Vorschau.png
