+++ dtos/__init__.py
@@ -1,0 +1,15 @@
+from dtos.catalogue_dto import CatalogueDto
+from dtos.found_task_dto import FoundTaskDto
+from dtos.subfield_dto import SubfieldDto
+from dtos.task_dto import TaskDto
+from dtos.field_dto import FieldDto
+from dtos.credentials_dto import ConfigDto
+
+__all__ = [
+	"CatalogueDto",
+	"FoundTaskDto",
+	"SubfieldDto",
+	"TaskDto",
+	"FieldDto",
+    "ConfigDto",
+]
