+++ dtos/field_dto.py
@@ -1,5 +1,5 @@
 from typing import Any
-from Dtos.subfield_dto import SubfieldDto
+from dtos.subfield_dto import SubfieldDto
 
 class FieldDto:
     def __init__(self,
