+++ program_logic/json_loader.py
@@ -1,5 +1,5 @@
 import json
-from Dtos import *
+from dtos import *
 from program_logic.path import Path
 
 class JsonLoader:
